requirements.txt
# ========================================
# CTI-NLP Enhanced Analyzer - Dependencies
# ========================================

# Core Framework
Flask==3.0.0
Flask-CORS==4.0.0
flask-socketio==5.3.5
python-socketio==5.10.0

# Machine Learning & Data Science
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2

# Optional: Intel GPU Acceleration (Remove if not needed)
#intel-extension-for-sklearn>=2023.1.1
#scikit-learn-intelex>=2023.1.1

# HTTP & API Requests
requests==2.31.0
urllib3==2.1.0

# Environment & Configuration
python-dotenv==1.0.0

# System Monitoring
psutil==5.9.6
watchdog==4.0.0

# IP Geolocation & Threat Intelligence
geoip2==4.7.0
maxminddb==2.5.1
ipaddress==1.0.23

# Security & Scanning
pefile==2023.2.7
python-magic==0.4.27
# hashlib and sqlite3 are built-in with Python

# Optional: Advanced Threat Intelligence
# abuseipdb-wrapper==0.2.1  # Requires API key
# shodan==1.31.0            # Requires API key

# Utilities
colorama==0.4.6
tqdm==4.66.1
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
celery==5.3.4
redis==5.0.1
